feat(core): Split TrustedKeyService into a read-only lookup and a main-only sync service (no-changelog)
Second step of the token-exchange/identity-substrate split (IAM-1181).
TrustedKeyService kept two lifecycles in one class: read-only lookups
(getByKidAndIss, hasSingleTrustedIssuer, isSsoIssuer, listAll,
listSources) that are safe on every instance type, and a write/refresh
lifecycle (config parsing, DB sync, JWKS refresh, leader/shutdown
hooks) that only ever runs on main.

Splits these into identity-substrate/services/trusted-key.service.ts
(read-only) and a new trusted-key-sync.service.ts (write/refresh).
Every dependent (identity-resolution.service.ts, token-exchange.service.ts,
trusted-key-source.controller.ts) now depends on the read-only class;
token-exchange.module.ts calls TrustedKeySyncService.initialize() and
registerSsoDerivedSource() for the write path.

No module-boundary change yet — both classes still live behind
token-exchange's main-only init(), so this stays behaviorally a no-op.
