+++ packages/cli/src/modules/identity-substrate/services/trusted-key-sync.service.ts
@@ -7,31 +7,27 @@ import type { EntityManager } from '@n8n/typeorm';
 import { In, Not } from '@n8n/typeorm';
 import { InstanceSettings } from 'n8n-core';
 import { UnexpectedError, UserError, jsonParse } from 'n8n-workflow';
-import type { KeyObject } from 'node:crypto';
 import { createHash, createPublicKey } from 'node:crypto';
 import { z } from 'zod';
 
-import { TrustedKeySourceEntity } from '@/modules/identity-substrate/database/entities/trusted-key-source.entity';
-import { TrustedKeyEntity } from '@/modules/identity-substrate/database/entities/trusted-key.entity';
-import { TrustedKeySourceRepository } from '@/modules/identity-substrate/database/repositories/trusted-key-source.repository';
-import { TrustedKeyRepository } from '@/modules/identity-substrate/database/repositories/trusted-key.repository';
-import { JwksResolverService } from '@/modules/identity-substrate/services/jwks-resolver';
-
-import { TokenExchangeConfig } from '../token-exchange.config';
+import { TokenExchangeConfig } from '@/modules/token-exchange/token-exchange.config';
 import type {
 	JwksKeySource,
 	JwtAlgorithm,
-	ResolvedTrustedKey,
 	StaticKeySource,
 	TrustedKeyData,
 	TrustedKeySource,
 	TrustedKeySourcePolicy,
-} from '../token-exchange.schemas';
+} from '@/modules/token-exchange/token-exchange.schemas';
 import {
-	TrustedKeyDataSchema,
 	TrustedKeySourcePolicySchema,
 	TrustedKeySourceSchema,
-} from '../token-exchange.schemas';
+} from '@/modules/token-exchange/token-exchange.schemas';
+
+import { JwksResolverService } from './jwks-resolver';
+import { TrustedKeySourceEntity } from '../database/entities/trusted-key-source.entity';
+import { TrustedKeyEntity } from '../database/entities/trusted-key.entity';
+import { TrustedKeySourceRepository } from '../database/repositories/trusted-key-source.repository';
 
 type AlgorithmFamily = 'RSA' | 'EC' | 'EdDSA';
 
@@ -54,7 +50,7 @@ const STATIC_SOURCE_ID = 'static';
 const REFRESH_POLL_INTERVAL_MS = 30 * Time.seconds.toMilliseconds;
 
 /**
- * Manages trusted public keys for JWT signature verification.
+ * Syncs and refreshes trusted public key sources into the database.
  *
  * Every instance resolves all configured key sources (env-var static keys,
  * JWKS endpoints) and writes them to the database at startup. Concurrent
@@ -65,28 +61,21 @@ const REFRESH_POLL_INTERVAL_MS = 30 * Time.seconds.toMilliseconds;
  * keys available in the database, avoiding a multi-main startup race where
  * a follower could previously verify against an empty table.
  *
- * All instances read keys from the database on every lookup, so multi-instance
- * consistency is preserved. A local crypto-primitive cache avoids repeated
- * `createPublicKey()` calls when the underlying key material has not changed.
+ * Only ever instantiated where it matters (`main`) — see
+ * `TrustedKeyService` for the read-only lookups every instance type uses.
  */
 @Service()
-export class TrustedKeyService {
+export class TrustedKeySyncService {
 	private readonly logger: Logger;
 
 	private refreshInterval: NodeJS.Timeout | undefined;
 
 	private isShuttingDown = false;
 
-	private readonly cryptoCache = new Map<
-		string,
-		{ keyMaterialHash: string; cryptoKey: KeyObject }
-	>();
-
 	constructor(
 		logger: Logger,
 		private readonly config: TokenExchangeConfig,
 		private readonly trustedKeySourceRepository: TrustedKeySourceRepository,
-		private readonly trustedKeyRepository: TrustedKeyRepository,
 		private readonly instanceSettings: InstanceSettings,
 		private readonly dbLockService: DbLockService,
 		private readonly jwksResolverService: JwksResolverService,
@@ -151,63 +140,6 @@ export class TrustedKeyService {
 		this.stopRefresh();
 	}
 
-	// ─── Public read path ──────────────────────────────────────────────
-
-	/**
-	 * Look up a resolved trusted key by its `kid` and `issuer`.
-	 *
-	 * Queries the database on every call (no stale reads). The local
-	 * crypto cache avoids repeated `createPublicKey()` calls when the
-	 * key material has not changed.
-	 */
-	async getByKidAndIss(kid: string, issuer: string): Promise<ResolvedTrustedKey | undefined> {
-		const entities = await this.trustedKeyRepository.findAllByKid(kid);
-		if (entities.length === 0) return undefined;
-
-		for (const entity of entities) {
-			let data: TrustedKeyData;
-			try {
-				const parsed = TrustedKeyDataSchema.safeParse(JSON.parse(entity.data));
-				if (!parsed.success) {
-					this.logger.warn('Skipping corrupted trusted key entity', {
-						kid,
-						sourceId: entity.sourceId,
-						error: parsed.error.message,
-					});
-					continue;
-				}
-				data = parsed.data;
-			} catch {
-				this.logger.warn('Skipping corrupted trusted key entity', {
-					kid,
-					sourceId: entity.sourceId,
-					error: 'invalid JSON',
-				});
-				continue;
-			}
-
-			if (data.issuer !== issuer) continue;
-
-			const cryptoKey = this.resolveCryptoKey(`${entity.sourceId}:${kid}`, data.keyMaterial);
-			if (!cryptoKey) continue;
-
-			return {
-				sourceId: entity.sourceId,
-				kid,
-				algorithms: data.algorithms,
-				key: cryptoKey,
-				issuer: data.issuer,
-				expectedAudience: data.expectedAudience,
-				inboundAudiences: data.inboundAudiences,
-				allowedRoles: data.allowedRoles,
-				requireVerifiedEmail: data.requireVerifiedEmail ?? true,
-				subjectClaim: data.subjectClaim ?? 'sub',
-			};
-		}
-
-		return undefined;
-	}
-
 	// ─── Public admin/diagnostic methods ───────────────────────────────
 
 	/**
@@ -220,14 +152,6 @@ export class TrustedKeyService {
 			throw new UnexpectedError(`Trusted key source not found: ${sourceId}`);
 		}
 		await this.refreshSourceInternal(source);
-	}
-
-	async listAll(): Promise<TrustedKeyEntity[]> {
-		return await this.trustedKeyRepository.find();
-	}
-
-	async listSources(): Promise<TrustedKeySourceEntity[]> {
-		return await this.trustedKeySourceRepository.find();
 	}
 
 	/**
@@ -292,46 +216,6 @@ export class TrustedKeyService {
 		}
 
 		return result.data;
-	}
-
-	async hasSingleTrustedIssuer(): Promise<boolean> {
-		const sources = await this.listAll();
-		const issuers = new Set<string>();
-
-		sources.forEach((entity) => {
-			try {
-				const parsed = TrustedKeyDataSchema.safeParse(JSON.parse(entity.data));
-				if (!parsed.success) {
-					this.logger.warn('Skipping corrupted trusted key entity', {
-						kid: entity.kid,
-						sourceId: entity.sourceId,
-						error: parsed.error.message,
-					});
-					return;
-				}
-				issuers.add(parsed.data.issuer);
-			} catch {
-				this.logger.warn('Skipping corrupted trusted key entity', {
-					kid: entity.kid,
-					sourceId: entity.sourceId,
-					error: 'invalid JSON',
-				});
-			}
-		});
-		return issuers.size === 1;
-	}
-
-	/**
-	 * Whether `issuer` is the instance's configured SSO provider — i.e. it
-	 * matches a `sso-derived` trusted key source (see
-	 * `registerSsoDerivedSource`). Indexed read only: no network, no crypto,
-	 * safe to call on every access.
-	 */
-	async isSsoIssuer(issuer: string): Promise<boolean> {
-		const source = await this.trustedKeySourceRepository.findOne({
-			where: { issuer, managedBy: 'sso-derived' },
-		});
-		return source !== null;
 	}
 
 	// ─── Private: source sync ──────────────────────────────────────────
@@ -799,29 +683,6 @@ export class TrustedKeyService {
 			throw new UnexpectedError(
 				`Trusted key "${kid}": key type "${keyType}" does not match algorithm family "${family}"`,
 			);
-		}
-	}
-
-	// ─── Private: crypto cache ─────────────────────────────────────────
-
-	private resolveCryptoKey(cacheKey: string, keyMaterial: string): KeyObject | undefined {
-		const hash = createHash('sha256').update(keyMaterial).digest('hex');
-		const cached = this.cryptoCache.get(cacheKey);
-
-		if (cached && cached.keyMaterialHash === hash) {
-			return cached.cryptoKey;
-		}
-
-		try {
-			const cryptoKey = createPublicKey(keyMaterial);
-			this.cryptoCache.set(cacheKey, { keyMaterialHash: hash, cryptoKey });
-			return cryptoKey;
-		} catch (error) {
-			this.logger.warn('Failed to parse key material from DB', {
-				cacheKey,
-				error: error instanceof Error ? error.message : String(error),
-			});
-			return undefined;
 		}
 	}
 }
