+++ packages/cli/src/modules/token-exchange/services/__tests__/token-exchange.service.test.ts
@@ -7,6 +7,7 @@ import { mock } from 'vitest-mock-extended';
 import { AuthError } from '@/errors/response-errors/auth.error';
 import { BadRequestError } from '@/errors/response-errors/bad-request.error';
 import type { JtiStoreService } from '@/modules/identity-substrate/services/jti-store.service';
+import type { TrustedKeyService } from '@/modules/identity-substrate/services/trusted-key.service';
 import type { JwtService } from '@/services/jwt.service';
 
 import type { InboundAudienceService } from '../../context-establishment-hooks/inbound-audience.service';
@@ -15,7 +16,6 @@ import { TokenExchangeAuthError } from '../../token-exchange.errors';
 import type { ResolvedTrustedKey } from '../../token-exchange.schemas';
 import type { IdentityResolutionService } from '../identity-resolution.service';
 import { TokenExchangeService } from '../token-exchange.service';
-import type { TrustedKeyService } from '../trusted-key.service';
 
 const logger = mock<Logger>({ scoped: vi.fn().mockReturnThis() });
 const trustedKeyStore = mock<TrustedKeyService>();
